app/requirements.txt
flask
mysql-connector-python
Flask-SQLAlchemy
python-dotenv